+++ eslint.config.cjs
@@ -11,6 +11,18 @@ module.exports = [
             parserOptions: {
                 ecmaVersion: 2022,
                 sourceType: 'module'
+            },
+            globals: {
+                setTimeout: 'readonly',
+                clearInterval: 'readonly',
+                setInterval: 'readonly',
+                clearTimeout: 'readonly',
+                Buffer: 'readonly',
+                require: 'readonly',
+                __dirname: 'readonly',
+                process: 'readonly',
+                console: 'readonly',
+                Thenable: 'readonly'
             }
         },
         plugins: {
@@ -27,7 +39,40 @@ module.exports = [
             curly: 'warn',
             eqeqeq: 'warn',
             'no-throw-literal': 'warn',
-            semi: 'warn'
+            semi: 'warn',
+            'no-unused-vars': 'off'
+        }
+    },
+    {
+        files: ['**/test/**/*.ts'],
+        languageOptions: {
+            parser: tsparser,
+            parserOptions: {
+                ecmaVersion: 2022,
+                sourceType: 'module'
+            },
+            globals: {
+                suite: 'readonly',
+                test: 'readonly',
+                setup: 'readonly',
+                teardown: 'readonly',
+                suiteSetup: 'readonly',
+                suiteTeardown: 'readonly',
+                setTimeout: 'readonly',
+                Buffer: 'readonly',
+                require: 'readonly',
+                __dirname: 'readonly',
+                process: 'readonly',
+                console: 'readonly',
+                Thenable: 'readonly'
+            }
+        },
+        plugins: {
+            '@typescript-eslint': tseslint
+        },
+        rules: {
+            '@typescript-eslint/naming-convention': 'off',
+            'no-unused-vars': 'off'
         }
     }
 ];